feat: shit form without refactor bc i have no time
