+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/capability/_components/grant-namespace-permission-form-utils.ts
