+++ apps/torus-portal/src/app/(pages)/permissions/create-permission/emission/_components/permission-success-dialog.tsx
